Added King Air mission analysis anddesign  optimization
